FIX CSP ERROR: Add Content Security Policy for Google Maps
REAL PROBLEM FOUND:
The issue is NOT the API key, but Content Security Policy blocking Google Maps!

Console Error:
- Refused to connect to maps.googleapis.com
- Violates Content Security Policy directive connect-src self

Solution:
Added CSP meta tag in public/index.html to allow:
- script-src: Google Maps scripts
- style-src: Google Fonts
- img-src: Map images
- connect-src: Google Maps API connections
- frame-src: Google iframes

This is the actual cause of the map not loading.
No need to change API key - it was working all along!